// front/src/types/industry.ts
export type IndustryT = {
    number: string;
    title: string;
    description: string;
    photo1: string;
    photo2: string;
    item1title: string;
    item1text: string;
    item2title: string;
    item2text: string;
    item3title: string;
    item3text: string;
    item4title: string;
    item4text: string;
    item5title: string;
    item5text: string;
}